pyproject.toml
[project]
name = "home-ops-cli"
version = "0.2.1"
description = ""
requires-python = ">=3.13"
dependencies = [
    "hvac==2.4.0",
    "ruamel.yaml==0.18.16",
    "kubernetes-asyncio==33.3.0",
    "kr8s==0.20.14",
    "aiohttp==3.13.2",
    "aiofiles==25.1.0",
    "aiogzip==1.1.0",
    "textual==6.7.1",
    "types-hvac==2.4.0.20251115",
    "boto3==1.42.3",
    "boto3-stubs==1.42.3",
    "mypy-boto3-s3==1.42.3",
    "inquirer==3.4.1",
    "pyusb==1.3.1",
    "textual-speedups==0.2.1",
    "typer==0.20.0",
    "kubevirt-py==1.7.0"
]

[project.scripts]
home-ops-cli = "home_ops_cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
